adding changes to config template service for other params
