+++ Core/resources-test/com/infiniteautomation/mango/spring/service/configTemplateTestData.csv
