+++ Core/resources-test/com/infiniteautomation/mango/spring/service/roles.mustache
@@ -1,0 +1,26 @@
+[
+  {
+  "xid": "ROLE_ADMIN",
+  "inherited": [],
+  "name": "Admin"
+  },
+  {
+  "xid": "ROLE_MANAGER",
+  "inherited": [],
+  "name": "Manager"
+  },
+  {{#.}}
+    {
+      "xid": "ROLE_SITE_{{site}}",
+      "inherited": [],
+      "name": "Role for {{site}}"
+    },
+    {{#dataHalls}}
+      {
+      "xid": "ROLE_DATAH_{{site}}_{{dataHall}}",
+      "inherited": [],
+      "name": "Role for site {{site}}, hall {{dataHall}}"
+      },
+    {{/dataHalls}}
+  {{/.}}
+  ]
