--- Core/resources/template.html
@@ -1,18 +1,0 @@
-template on ma-core-public core resources
-<!--
-  ~  Copyright (C) 2023 RadixIot LLC. All rights reserved.
-  -->
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Title</title>
-</head>
-<body>
-test3
-{{#.}}
-keylabel:{{key}} : valuelabel:{{value}}
-{{/.}}
-</body>
-</html>
